Update mass_scraping_HR.py
Added code at the end (row 190) to remove csv file that was created during scraping. This will change the output of xlsx file, because it will create new csv file instead of appending to the previous one and making wrong xlsx file at the end with info that is not correct (to many articles for example).